+++ src/com/kaylerrenslow/armaplugin/lang/PreprocessorColorizerAnnotator.java
@@ -1,17 +1,18 @@
-package com.kaylerrenslow.armaplugin.lang.header.psi.codestyle;
+package com.kaylerrenslow.armaplugin.lang;
 
 import com.intellij.lang.annotation.Annotation;
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.Annotator;
 import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.PsiElement;
+import com.kaylerrenslow.armaplugin.lang.header.psi.codestyle.HeaderSyntaxHighlighter;
 import org.jetbrains.annotations.NotNull;
 
 /**
  * @author Kayler
  * @since 09/11/2017
  */
-public class HeaderIncludeColorizerAnnotator implements Annotator {
+public class PreprocessorColorizerAnnotator implements Annotator {
 
 	@Override
 	public void annotate(@NotNull PsiElement element, @NotNull AnnotationHolder holder) {